Globe component and styles
