+++ src/components/styles/Globe.css
@@ -1,0 +1,22 @@
+.globe-div {
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+  width: 100%;
+  padding-bottom: 20rem;
+  padding-top: 5rem;
+}
+#globe {
+  border-radius: 50%;
+  width: 100rem;
+  height: 100rem;
+  background-color: transparent;
+  display: inline-block;
+  transform: scale(1);
+}
+
+#globe-image {
+  max-width: 100%;
+  max-height: 100%;
+  opacity: 1;
+}
