+++ src/components/styles/HomeContent.css
@@ -9,28 +9,6 @@
   width: 40rem;
   animation: transitionIn 1s;
 }
-.globe-div {
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  width: 100%;
-  padding-bottom: 20rem;
-  padding-top: 5rem;
-}
-#globe {
-  border-radius: 50%;
-  width: 100rem;
-  height: 100rem;
-  background-color: #002a5c;
-  display: inline-block;
-  transform: scale(1);
-}
-
-#globe-image {
-  max-width: 100%;
-  max-height: 100%;
-  opacity: 1;
-}
 
 @keyframes transitionIn {
   from {
